--- site/docs/benchmark.mdx
@@ -1,9 +1,0 @@
----
-sidebar_position: 5
----
-
-# Benchmark
-
-import BenchmarkResultCharts from '@site/src/components/BenchmarkResultCharts'
-
-<BenchmarkResultCharts/>
