+++ site/package.json
@@ -20,16 +20,17 @@
     "@docusaurus/preset-classic": "2.2.0",
     "@easyops-cn/docusaurus-search-local": "^0.21.4",
     "@mdx-js/react": "^1.6.21",
-    "chart.js": "^2.9.4",
+    "chart.js": "^3.9.1",
     "clsx": "^1.1.1",
     "prism-react-renderer": "^1.2.1",
     "react": "^17.0.1",
-    "react-chartjs-2": "^2.11.2",
+    "react-chartjs-2": "^4.3.1",
     "react-dom": "^17.0.1"
   },
   "devDependencies": {
     "@docusaurus/module-type-aliases": "2.2.0",
     "@tsconfig/docusaurus": "^1.0.4",
+    "@types/chart.js": "^2.9.37",
     "typescript": "^4.5.2"
   },
   "browserslist": {
